Moved SocialMediaService helper to lib/utils
